refactoring test files and removing flaky stress test
